+++ src/lib/lforms/lhc-form-data.ts
@@ -1967,6 +1967,11 @@ export default class LhcFormData {
     var readerMsg = 'Added ' + this.itemDescription(item);
     this._actionLogs.push(readerMsg);
 
+    // run FHIRPATH expression when a new item is added
+    if (LForms.FHIR && this._hasResponsiveExpr) {
+      this._expressionProcessor.runCalculations(false).then(()=>{});
+    }
+
     return newItem;
   }
 
@@ -2194,6 +2199,12 @@ export default class LhcFormData {
     this._resetInternalData();
     var readerMsg = 'Removed ' + this.itemDescription(item);
     this._actionLogs.push(readerMsg);
+
+    // run FHIRPATH expression when a new item is removed
+    if (LForms.FHIR && this._hasResponsiveExpr) {
+      this._expressionProcessor.runCalculations(false).then(()=>{});
+    }
+    
   }